ai-model-main/requirement.txt
langchain
wikipedia
langchain-community
langchain-openai
langchain-anthropic
python-dotenv
pydantic
DuckDuckgo-search